+++ CHANGES.md
@@ -1,1 +1,2 @@
+- Switch "Start Next" button to a search button if no orders are shown on the orders page
 - Fix not allocating reagents for orders with some reagents provided